+++ ChangeLog.md
@@ -5,5 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+-  NEW :  création du module CliEnjoyHolidays - *2024-04-08* - 1.1
+
 
 -  NEW :  initialisation du module CliEnjoyHolidays - *2024-04-08* - 1.0